Add webserver based epaper demo
Plenty of room for improvement but now we can set the name on the badge.
Starts an AP with a simple server and prints instructions on the epaper.
The user connects to the AP and enters the given URL to set the name.

Badge sleeps until the send button is pressed to conserve batteries.

Signed-off-by: Michael Welling <mwelling@ieee.org>