display space character in first or last name
This change fixes an error in the logic of
G_display_string_at() which is used to display
the first and last name in a large font.

The space character is the first element in
G_FreeSans24pt7b.Glyphs. A bug in the logic
of G_display_string_at() was causing it to
skip displaying the first element.